+++ services/analyzer/sequelize-tables-analyzer.js
@@ -1,9 +1,15 @@
 const P = require('bluebird');
 const _ = require('lodash');
+const { plural, singular } = require('pluralize');
 const ColumnTypeGetter = require('./sequelize-column-type-getter');
 const TableConstraintsGetter = require('./sequelize-table-constraints-getter');
 const { DatabaseAnalyzerError } = require('../../utils/errors');
 const { terminate } = require('../../utils/terminator');
+
+const ASSOCIATION_TYPE_BELONGS_TO = 'belongsTo';
+const ASSOCIATION_TYPE_BELONGS_TO_MANY = 'belongsToMany';
+const ASSOCIATION_TYPE_HAS_MANY = 'hasMany';
+const ASSOCIATION_TYPE_HAS_ONE = 'hasOne';
 
 let queryInterface;
 let tableConstraintsGetter;
@@ -79,52 +85,159 @@ function hasIdColumn(fields, primaryKeys) {
     || _.includes(primaryKeys, 'id');
 }
 
+// NOTICE: Check the foreign key's reference unicity
+function checkUnicity(primaryKeys, uniqueIndexes, columnName) {
+  const isUnique = uniqueIndexes !== null
+    && uniqueIndexes.find((indexColumnName) =>
+      indexColumnName.length === 1 && indexColumnName.includes(columnName));
+
+  const isPrimary = _.isEqual([columnName], primaryKeys);
+  return { isPrimary, isUnique };
+}
+
+// NOTICE: Format the references depending on the type of the association
+function createReference(tableName, association, foreignKey, manyToManyForeignKey) {
+  const reference = {
+    foreignKey: foreignKey.columnName,
+    foreignKeyName: _.camelCase(foreignKey.columnName),
+    association,
+  };
+
+  if (association === ASSOCIATION_TYPE_BELONGS_TO) {
+    reference.ref = foreignKey.foreignTableName;
+    reference.as = formatAliasName(foreignKey.columnName);
+  } else if (association === ASSOCIATION_TYPE_BELONGS_TO_MANY) {
+    reference.ref = manyToManyForeignKey.foreignTableName;
+    reference.otherKey = manyToManyForeignKey.columnName;
+    reference.junctionTable = foreignKey.tableName;
+  } else {
+    reference.ref = foreignKey.tableName;
+
+    const formater = association === ASSOCIATION_TYPE_HAS_MANY ? plural : singular;
+    const prefix = (singular(tableName) === formatAliasName(reference.foreignKeyName))
+      ? ''
+      : `${formatAliasName(reference.foreignKeyName)}_`;
+
+    reference.as = _.camelCase(formater(`${prefix}${foreignKey.tableName}`));
+  }
+
+  // NOTICE: If the foreign key name and alias are the same, Sequelize will crash, we need
+  //         to handle this specific scenario generating a different foreign key name.
+  if (reference.foreignKeyName && reference.foreignKeyName === reference.as) {
+    reference.foreignKeyName = `${reference.foreignKeyName}Key`;
+  }
+
+  if (foreignKey.foreignColumnName !== 'id') {
+    reference.targetKey = foreignKey.foreignColumnName;
+  }
+
+  return reference;
+}
+
 async function analyzeTable(table, config) {
   const schema = await analyzeFields(table, config);
 
   return {
     schema,
-    foreignKeys: await tableConstraintsGetter.perform(table),
+    constraints: await tableConstraintsGetter.perform(table),
     primaryKeys: await analyzePrimaryKeys(schema),
   };
 }
 
+// NOTICE: Use the foreign key and reference properties to determine the associations
+//         and push them as references of the table.
+function defineAssociationType(databaseSchema) {
+  Object.values(databaseSchema).forEach((table) => {
+    const { constraints, primaryKeys } = table;
+    constraints.forEach((constraint) => {
+      const { columnName } = constraint;
+      const uniqueIndexes = constraint.uniqueIndexes || null;
+
+      if (constraint.columnType === 'FOREIGN KEY') {
+        const isInCompositeKey = primaryKeys
+          && primaryKeys.length > 1
+          && primaryKeys.includes(columnName);
+
+        const { isPrimary, isUnique } = checkUnicity(primaryKeys, uniqueIndexes, columnName);
+
+        const referenceTableName = constraint.foreignTableName;
+        const referenceColumnName = constraint.foreignColumnName;
+        let isManyToMany = false;
+
+        if (isInCompositeKey) {
+          // Check if the foreignKey is in a composite primary key
+          primaryKeys.forEach((primaryKey) => {
+            if (primaryKey === constraint.columnName) {
+              const manyToManyKeys = _.filter(constraints, (otherKey) =>
+                otherKey.columnName !== constraint.columnName
+                  && otherKey.columnType === 'FOREIGN KEY' && primaryKeys.includes(otherKey.columnName)) || [];
+
+              manyToManyKeys.forEach((manyToManyKey) => {
+                databaseSchema[referenceTableName].references.push(
+                  createReference(
+                    referenceTableName,
+                    ASSOCIATION_TYPE_BELONGS_TO_MANY,
+                    constraint,
+                    manyToManyKey,
+                  ),
+                );
+              });
+              isManyToMany = manyToManyKeys !== [];
+            }
+          });
+        }
+        if (!isManyToMany) {
+          const referencePrimaryKeys = databaseSchema[referenceTableName].primaryKeys;
+          const referenceUniqueConstraint = databaseSchema[referenceTableName].constraints
+            .find(({ columnType }) => columnType === 'UNIQUE');
+          const referenceUniqueIndexes = referenceUniqueConstraint
+            ? referenceUniqueConstraint.uniqueIndexes
+            : null;
+          const referenceUnicity = checkUnicity(
+            referencePrimaryKeys,
+            referenceUniqueIndexes,
+            referenceColumnName,
+          );
+
+          if (referenceUnicity.isPrimary || referenceUnicity.isUnique) {
+            table.references.push(
+              createReference(
+                null,
+                ASSOCIATION_TYPE_BELONGS_TO,
+                constraint,
+              ),
+            );
+          }
+          databaseSchema[referenceTableName].references.push(
+            createReference(
+              referenceTableName,
+              (isPrimary || isUnique) ? ASSOCIATION_TYPE_HAS_ONE : ASSOCIATION_TYPE_HAS_MANY,
+              constraint,
+            ),
+          );
+        }
+      }
+    });
+  });
+}
+
 async function createTableSchema({
   schema,
-  foreignKeys,
+  constraints,
   primaryKeys,
+  references,
 }, tableName) {
   const fields = [];
-  const references = [];
 
   await P.each(Object.keys(schema), async (columnName) => {
     const columnInfo = schema[columnName];
     const type = await columnTypeGetter.perform(columnInfo, columnName, tableName);
-    const foreignKey = _.find(foreignKeys, { column_name: columnName });
+    const foreignKey = _.find(constraints, { columnName, columnType: 'FOREIGN KEY' });
+    const isValidField = type && (!foreignKey
+      || !foreignKey.foreignTableName
+      || !foreignKey.columnName || columnInfo.primaryKey);
 
-    if (foreignKey
-      && foreignKey.foreign_table_name
-      && foreignKey.column_name
-      && !columnInfo.primaryKey) {
-      const reference = {
-        ref: foreignKey.foreign_table_name,
-        foreignKey: foreignKey.column_name,
-        foreignKeyName: _.camelCase(foreignKey.column_name),
-        as: formatAliasName(foreignKey.column_name),
-      };
-
-      // NOTICE: If the foreign key name and alias are the same, Sequelize will crash, we need
-      //         to handle this specific scenario generating a different foreign key name.
-      if (reference.foreignKeyName === reference.as) {
-        reference.foreignKeyName = `${reference.foreignKeyName}Key`;
-      }
-
-      if (foreignKey.foreign_column_name !== 'id') {
-        reference.targetKey = foreignKey.foreign_column_name;
-      }
-
-      references.push(reference);
-    } else if (type) {
+    if (isValidField) {
       // NOTICE: If the column is of integer type, named "id" and primary, Sequelize will
       //         handle it automatically without necessary declaration.
       if (!(columnName === 'id' && type === 'INTEGER' && columnInfo.primaryKey)) {
@@ -160,7 +273,7 @@ async function createTableSchema({
 
   return {
     fields,
-    references,
+    references: _.sortBy(references, 'association'),
     primaryKeys,
     options,
   };
@@ -192,11 +305,24 @@ async function analyzeSequelizeTables(databaseConnection, config, allowWarning) 
   }
 
   // Build the db schema.
+  const databaseSchema = {};
   const tableNames = await showAllTables(databaseConnection, config.dbSchema);
 
   await P.each(tableNames, async (tableName) => {
-    const tableAnalysis = await analyzeTable(tableName, config);
-    schemaAllTables[tableName] = await createTableSchema(tableAnalysis, tableName);
+    const { schema, constraints, primaryKeys } = await analyzeTable(tableName, config);
+    databaseSchema[tableName] = {
+      schema,
+      constraints,
+      primaryKeys,
+      references: [],
+    };
+  });
+
+  // Fill the references field for each table schema
+  defineAssociationType(databaseSchema);
+
+  await P.each(tableNames, async (tableName) => {
+    schemaAllTables[tableName] = await createTableSchema(databaseSchema[tableName], tableName);
   });
 
   if (_.isEmpty(schemaAllTables)) {