+++ test/fixtures/mysql/users.sql
@@ -1,0 +1,6 @@
+CREATE TABLE users (
+  id INT AUTO_INCREMENT PRIMARY KEY,
+  username VARCHAR(25) NOT NULL,
+  enabled BIT(1) DEFAULT 1,
+  last_login DATE NOT NULL
+);